Update everest-sqlite to 0.1.4, libevse-security to 0.9.9 and libwebsockets to 4.4.1 (#1143)
* Update everest-sqlite to 0.1.4
* Adapt database stub to fixed signature of column_int64 function
* Update libevse-security to 0.9.9
* Bump libwebsockets to v4.4.1
* Bump version to 0.30.2

---------

+++ tests/lib/ocpp/v16/database_stub.hpp
@@ -81,7 +81,7 @@ struct SQLiteStatementTest : public StatementInterface {
     virtual int column_int(const int idx) {
         return 0;
     }
-    virtual int64_t column_int64(const int64_t idx) {
+    virtual int64_t column_int64(const int idx) {
         return 0;
     }
     virtual double column_double(const int idx) {
